skip workaroubnd if no config

+++ ckanext/clamav/adapters.py
@@ -50,6 +50,15 @@ class CustomClamdNetworkSocket(ClamdNetworkSocket):
           - BufferTooLongError: if the buffer size exceeds clamd limits
           - ConnectionError: in case of communication problem
         """
+        
+        # Define writable directory and dummy file with random string
+        write_dir: str = tk.config.get(c.CLAMAV_CONF_WRITE_DIR, None)
+        if write_dir:
+            random_str = ''.join(random.choice(string.ascii_letters) for i in range(8))
+            dummy_file = f'{write_dir}/{random_str}'
+            print('write dir and dummy file:')
+            print(write_dir)
+            print(dummy_file)
 
         try:
             self._init_socket()
@@ -59,23 +68,16 @@ class CustomClamdNetworkSocket(ClamdNetworkSocket):
 
             chunk = buff.read(max_chunk_size)
 
-            # Define writable directory and dummy file with random string
-            write_dir: str = tk.config.get(c.CLAMAV_CONF_WRITE_DIR)
-            random_str = ''.join(random.choice(string.ascii_letters) for i in range(8))
-            dummy_file = f'{write_dir}/{random_str}'
-            print('write dir and dummy file:')
-            print(write_dir)
-            print(dummy_file)
-
             while chunk:
                 size = struct.pack(b'!L', len(chunk))
                 self.clamd_socket.send(size + chunk)
                 chunk = buff.read(max_chunk_size)
                 
-                # Generate dummy file
-                f = open(dummy_file, 'w')
-                f.write(random_str)
-                f.close()
+                # Generate temporary dummy file
+                if write_dir:
+                    f = open(dummy_file, 'w')
+                    f.write(random_str)
+                    f.close()
 
             self.clamd_socket.send(struct.pack(b'!L', 0))
 
@@ -89,6 +91,9 @@ class CustomClamdNetworkSocket(ClamdNetworkSocket):
                 return {filename: (status, reason)}
         
         finally:
-            os.remove(dummy_file)
+            # Remove temporary dummy file
+            if write_dir:
+                os.remove(dummy_file)
+
             self._close_socket()
